refactor(testing): apply architecture-review findings and restructure cert tests
Review fixes: one shared LineQueue replaces three line-buffer copies; PICS
unavailability is a typed condition, other pics errors rethrow; CertNodeApi
drops cluster-specific members in favor of a fabricFiltered read option;
explicit barrel exports; declaration-time step validation; device exits
observed outside a step now fail the run; default device flavor is matterjs.

Restructuring: framework self-tests move to test/cert-framework with
FRAMEWORK-* ids so test/cert holds only official TCs; the cert framework
unit tests move to support/chip-testing as @matter/testing consumers
(ESM-only by design); step-boundary banners are injected into all followed
logs; the test-plan feedback log moves out of the repo; evidence artifacts
carry a .zip name.
