+++ packages/testing/src/chip/cert/cert-test.ts
@@ -9,7 +9,7 @@ import { BaseTest } from "../../device/test.js";
 import type { Container } from "../../docker/container.js";
 import { TestDescriptor, TestFileDescriptor } from "../../test-descriptor.js";
 import { PicsExpression } from "../pics/expression.js";
-import type { PicsFile } from "../pics/file.js";
+import { PicsUnavailableError, type PicsFile } from "../pics/file.js";
 import {
     CertDevice,
     CertStepContext,
@@ -18,6 +18,7 @@ import {
     DeviceExitInfo,
     DeviceFlavor,
     StepRecorder,
+    StepVerdict,
 } from "./cert-context.js";
 
 const inertRecorder: StepRecorder = {
@@ -62,6 +63,7 @@ export class CertTest extends BaseTest {
         const picsFile = resolvePicsFile(subject);
         const deviceExitWatch = watchDeviceExits(devices, recorder);
         const flavor = currentFlavor(devices);
+        const tc = this.#definition.tc;
 
         let aborted = false;
         let failure: unknown;
@@ -70,38 +72,31 @@ export class CertTest extends BaseTest {
         try {
             for (const stepDef of this.#definition.steps) {
                 if (aborted) {
+                    announceStepEnd(cx, tc, stepDef, "aborted");
                     recorder.endStep(stepDef, "aborted");
                     continue;
                 }
 
                 try {
-                    if (stepDef.flavors?.length === 0) {
-                        throw new Error(
-                            `Step ${stepDef.number} declares an empty "flavors" list, which would skip it on ` +
-                                "every flavor — omit the option instead if that's genuinely intended",
-                        );
-                    }
-
                     if (stepDef.flavors && flavor !== undefined && !stepDef.flavors.includes(flavor)) {
+                        announceStepEnd(cx, tc, stepDef, "skipped");
                         recorder.endStep(stepDef, "skipped", `unsupported on device flavor "${flavor}"`);
                         continue;
                     }
 
                     if (!stepPicsMet(stepDef, picsFile)) {
+                        announceStepEnd(cx, tc, stepDef, "skipped");
                         recorder.endStep(stepDef, "skipped", `PICS "${stepDef.pics}" not met`);
                         continue;
                     }
 
                     step(`Test Step ${stepDef.number}: ${stepDef.text}`);
+                    announceStepStart(cx, tc, stepDef);
                     recorder.beginStep(stepDef);
 
-                    await raceAgainstDeviceExit(
-                        stepDef.run(cx),
-                        deviceExitWatch.exit,
-                        this.#definition.tc,
-                        stepDef.number,
-                    );
+                    await raceAgainstDeviceExit(stepDef.run(cx), deviceExitWatch.exit, tc, stepDef.number);
                 } catch (e) {
+                    announceStepEnd(cx, tc, stepDef, "fail");
                     recorder.endStep(stepDef, "fail");
                     aborted = true;
                     failed = true;
@@ -109,6 +104,7 @@ export class CertTest extends BaseTest {
                     continue;
                 }
 
+                announceStepEnd(cx, tc, stepDef, "pass");
                 recorder.endStep(stepDef, "pass");
             }
         } finally {
@@ -136,6 +132,13 @@ export class CertTest extends BaseTest {
         if (failed) {
             throw failure;
         }
+
+        const exited = deviceExitWatch.observed;
+        if (exited !== undefined) {
+            throw new Error(
+                `A cert-test device exited unexpectedly (code ${exited.code}, signal ${exited.signal}) during the run`,
+            );
+        }
     }
 
     /**
@@ -161,14 +164,18 @@ export class CertTest extends BaseTest {
 
 /**
  * {@link Subject.pics} is typed as a required {@link PicsFile} but its real implementations can throw
- * (e.g. the harness has no active PICS file yet) rather than return one. Failing to obtain a PICS file
- * means no PICS gating is active, not that the test failed — every step's PICS is then treated as met.
+ * {@link PicsUnavailableError} (e.g. the harness has no active PICS file yet) rather than return one.
+ * That specific condition means no PICS gating is active, not that the test failed — every step's
+ * PICS is then treated as met. Any other error from the accessor is a real failure and propagates.
  */
 function resolvePicsFile(subject: Subject): PicsFile | undefined {
     try {
         return subject.pics ?? undefined;
-    } catch {
-        return undefined;
+    } catch (e) {
+        if (e instanceof PicsUnavailableError) {
+            return undefined;
+        }
+        throw e;
     }
 }
 
@@ -193,6 +200,37 @@ function stepPicsMet(stepDef: CertStepDefinition, picsFile: PicsFile | undefined
     return new PicsExpression(stepDef.pics).evaluate(picsFile);
 }
 
+const STEP_BANNER_RULE = "-".repeat(70);
+
+/**
+ * Injects a step-boundary banner (chip python/yaml style) into every device's and controller's log
+ * buffer, so a step's start/end is visible in the written evidence `.log` files at the
+ * chronologically right position. {@link LogFollower.annotate} flags each line synthetic, so a
+ * step's own `log.expect()` never matches a banner.
+ */
+function announceStep(cx: CertStepContext, lines: string[]): void {
+    for (const line of lines) {
+        for (const device of Object.values(cx.devices)) {
+            device.log.annotate(line);
+        }
+        for (const controller of Object.values(cx.controllers)) {
+            controller.log.annotate(line);
+        }
+    }
+}
+
+function announceStepStart(cx: CertStepContext, tc: string, stepDef: CertStepDefinition): void {
+    announceStep(cx, [STEP_BANNER_RULE, `${tc} — Test Step ${stepDef.number}: ${stepDef.text}`, STEP_BANNER_RULE]);
+}
+
+function announceStepEnd(cx: CertStepContext, tc: string, stepDef: CertStepDefinition, verdict: StepVerdict): void {
+    announceStep(cx, [
+        STEP_BANNER_RULE,
+        `${tc} — Test Step ${stepDef.number}: ${verdict.toUpperCase()}`,
+        STEP_BANNER_RULE,
+    ]);
+}
+
 /**
  * A {@link watchDeviceExits} subscription: `exit` resolves the same way every time (first device to
  * exit wins), but the reaction that reports it to `recorder` must be {@link disarm}ed once the run is
@@ -200,6 +238,13 @@ function stepPicsMet(stepDef: CertStepDefinition, picsFile: PicsFile | undefined
  */
 interface DeviceExitWatch {
     exit: Promise<DeviceExitInfo>;
+    /**
+     * The exit observed while the watch was armed, if any. A device exit can settle outside any
+     * step race (all steps skipped, or after the last step) — {@link CertTest.invoke} checks this
+     * after the step loop so such a run still rejects instead of reporting success while the
+     * evidence says fail.
+     */
+    readonly observed: DeviceExitInfo | undefined;
     /**
      * Drops the watch's reference to `recorder`. A matterjs device's `exit` never resolves, so
      * without this the reaction below stays attached to it for the process's lifetime, keeping
@@ -215,12 +260,17 @@ interface DeviceExitWatch {
  * the device crashed independently of anything the test asked it to do.
  */
 function watchDeviceExits(devices: Record<string, CertDevice>, recorder: StepRecorder): DeviceExitWatch {
+    let observed: DeviceExitInfo | undefined;
     let onExit: ((info: DeviceExitInfo) => void) | undefined = info => recorder.deviceExited?.(info);
     const exit = Promise.race(Object.values(devices).map(device => device.exit));
 
     void exit.then(info => {
+        if (onExit === undefined) {
+            return;
+        }
+        observed = info;
         try {
-            onExit?.(info);
+            onExit(info);
         } catch (e) {
             console.warn("Cert test deviceExited hook failed:", e);
         }
@@ -228,6 +278,9 @@ function watchDeviceExits(devices: Record<string, CertDevice>, recorder: StepRec
 
     return {
         exit,
+        get observed() {
+            return observed;
+        },
         disarm() {
             onExit = undefined;
         },
@@ -250,7 +303,7 @@ async function raceAgainstDeviceExit(
         // its eventual rejection so it can't surface as an unhandled rejection attributed to
         // whatever runs later, without trying to cancel the operation itself.
         void stepRun.catch(e => {
-            console.debug(`Cert test ${tc} step ${step}: orphaned step run settled after the device exit:`, e);
+            console.warn(`Cert test ${tc} step ${step}: orphaned step run settled after the device exit:`, e);
         });
         throw new Error("A cert-test device exited unexpectedly while a step was running");
     }
