+++ packages/testing/src/index.d.ts
@@ -11,5 +11,8 @@ export * from "./docker/index.js";
 export { afterRun } from "./mocha.js";
 export * from "./mocharc.cjs";
 export * from "./runner.js";
+/** @internal Test seam — not API. Needed to construct a {@link CertTest} directly in a unit test. */
+export type { TestFileDescriptor } from "./test-descriptor.js";
+export { LineQueue } from "./util/async.js";
 export * from "./util/heap.js";
 export * from "./util/wtf.js";
