+++ packages/testing/src/index.js
@@ -11,5 +11,6 @@ export * from "./docker/index.js";
 export { afterRun } from "./mocha.js";
 export * from "./mocharc.cjs";
 export * from "./runner.js";
+export { LineQueue } from "./util/async.js";
 export * from "./util/heap.js";
 export * from "./util/wtf.js";
