+++ support/chip-testing/package.json
@@ -28,7 +28,8 @@
         "test-app-cc-1": "matter-test --spec=test/app-cc/CC.1.test.ts --report",
         "test-app-cc-2": "matter-test --spec=test/app-cc/CC.2.test.ts --report",
         "test-icd": "matter-test --spec=test/icd/**/*.test.ts --report",
-        "test-cert": "MATTER_TEST_SHUTDOWN_TIMEOUT_MS=15000 matter-test --spec=test/cert/**/*.test.ts --report"
+        "test-cert": "MATTER_TEST_SHUTDOWN_TIMEOUT_MS=15000 matter-test --spec=test/cert/**/*.test.ts --report",
+        "test-cert-framework": "MATTER_TEST_SHUTDOWN_TIMEOUT_MS=15000 matter-test --spec=test/cert-framework/**/*.test.ts --report"
     },
     "dependencies": {
         "@matter/main": "*",
